增加spring mock测试

+++ spring/src/main/webapp/WEB-INF/sql/user.sql
@@ -22,7 +22,7 @@ SET FOREIGN_KEY_CHECKS = 0;
 -- ----------------------------
 DROP TABLE IF EXISTS `user`;
 CREATE TABLE `user` (
-  `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
+  `id` bigint(10) unsigned NOT NULL AUTO_INCREMENT,
   `name` varchar(64) NOT NULL DEFAULT '',
   `dept` varchar(254) NOT NULL DEFAULT '',
   `website` varchar(254) DEFAULT '',
